requirements.txt
absl-py==2.1.0
alembic==1.14.1
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
DataProperty==1.0.1
docker-pycreds==0.4.0
-e git+https://github.com/matolevi/energy-net.git@23452cb430d3f5b0a8f04e98346c543c92bd8a94#egg=energy_net
eval_type_backport==0.2.2
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.70.0
gymnasium==1.0.0
huggingface-hub==0.29.3
huggingface-sb3==3.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mbstrdecoder==1.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
optuna==4.2.1
packaging==24.2
pandas==2.0.3
pathvalidate==3.2.1
pettingzoo==1.24.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rl_zoo3==2.4.0
rsa==4.9
sb3_contrib==2.4.0
scipy==1.10.1
sentry-sdk==2.22.0
setproctitle==1.3.5
Shimmy==2.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.39
stable_baselines3==2.4.1
sympy==1.13.3
tabledata==1.3.3
tcolorpy==0.1.6
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.2.2
tqdm==4.67.1
typepy==1.3.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wandb==0.19.8
wasabi==1.1.3
Werkzeug==3.0.6
zipp==3.20.2
